requirements.txt
google-generativeai==0.8.4
chromadb==0.6.3
PyPDF2==3.0.1
python-docx==1.1.2
pandas==2.2.3
beautifulsoup4==4.12.3
rich==13.9.4
python-dotenv==1.0.1
tqdm==4.67.1